requirements.txt
requests
appdirs
aria2p
progress
psutil
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
js2py
python-dotenv
tenacity
python-magic
beautifulsoup4
Pyrogram
TgCrypto
psycopg2-binary
yt-dlp
lxml
telegraph
speedtest-cli
messages
aiohttp
feedparser
